requirements.txt
flask==2.0.1
flask-cors==3.0.10
schedule==1.1.0
python-dotenv==0.19.0
requests==2.26.0
pytrends==4.7.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
firebase-admin==5.0.0
google-generativeai==0.3.0
python-dateutil==2.8.2
aiohttp==3.8.1
beautifulsoup4==4.9.3
asyncio==3.4.3 